Add the matcher, premium backfill, and priority scoring
Joins the feeds onto the catalog. Entries the free repository does not
carry are brought in as premium rather than left as dangling slugs, so
every plugin a feed names is in the catalog: 10,756 of them.

A finding means the currently published version sits inside an affected
range, which is 1,294 of 46,402 ranges. Priority is scored separately
off the whole issue history, because a plugin with a long record of
holes is worth scanning whether or not today's release happens to be
clean.

Two performance and correctness fixes found by measuring rather than
guessing:

- rescore rewrote all 82k catalog rows every pass and spent two minutes
  in index maintenance. It now only touches rows whose values actually
  changed, which took it to 26 seconds and to zero writes on a steady
  week.
- priority_score and cvss_score were mysql FLOAT, which is single
  precision, so a score round-tripped lossily and never compared equal
  to itself. 268 rows were rewritten every pass regardless. Both are
  DOUBLE now.

+++ kpwpvs/models/software.py
@@ -21,7 +21,7 @@ from sqlalchemy import (
     Boolean,
     Date,
     DateTime,
-    Float,
+    Double,
     ForeignKey,
     Index,
     Integer,
@@ -126,7 +126,7 @@ class Software(TimestampMixin, Base):
     issue_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0, index=True)
     open_issue_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
     critical_issue_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    priority_score: Mapped[float] = mapped_column(Float, nullable=False, default=0.0)
+    priority_score: Mapped[float] = mapped_column(Double, nullable=False, default=0.0)
 
     # crawl bookkeeping, last_seen tells us when it fell out of the catalog
     first_seen: Mapped[datetime | None] = mapped_column(DateTime, nullable=True)
